value dependency
previous implementation had the `value` in a useEffect with other state
changes. I remove the uncessary `useEffect`, and set just one dependent
on `value`, so when `value` is null, the component can be reset. To
avoid passing `value` into the `List.tsx` component, a new ref was
created that kep track of the previously selected item, allowing the
component to know what the previuos itme was so it could avoid an
uncessary call to `action`.