double submit
While focused on a select item, when the user hit `Enter` the `onSubmit`
would fire switch on firefox, overwriting the `enter` event set on the
itme list. A boolean comparing the currnet name and project name was
inserted to return, thus allowing the `keydown` event on the item to
continue.